feat(aws.showRegion): "Show or Hide Regions"
- change `aws.showRegion` to support multiselect
    - allow showing and hiding in a single "Show or Hide Regions" command
- remove `aws.hideRegion`
