+++ .changes/next-release/Feature-bcef224b-f313-45fb-aa0f-efc458cf6016.json
@@ -1,0 +1,4 @@
+{
+	"type": "Feature",
+	"description": "\"Show or Hide Regions\" is now one command that supports multiple selections"
+}
