+++ README.quickstart.cloud9.md
@@ -91,8 +91,7 @@ The Toolkit provides commands (prefixed with `AWS:`) to the AWS Cloud9 _Go to An
 | `AWS: Deploy SAM Application`             | Deploys a local serverless application to an AWS account. For more information, see [Deploying a Serverless Application](https://docs.aws.amazon.com/cloud9/latest/user-guide/deploy-serverless-app.html) in the user guide.               |
 | `AWS: Edit SAM Debug Configuration`       | Shows a tool that helps you create, edit, run, and debug a SAM _launch config_ (`type:aws-sam`).                                                                                                                                           |
 | `AWS: Detect SAM CLI`                     | Checks whether the Toolkit can communicate correctly with the AWS SAM CLI that is installed.                                                                                                                                               |
-| `AWS: Hide region from the Explorer`      | Hides an AWS Region from the **AWS Explorer**.                                                                                                                                                                                             |
-| `AWS: Show region in the Explorer`        | Displays an AWS Region in the **AWS Explorer**.                                                                                                                                                                                            |
+| `AWS: Show or Hide Regions`               | Adds or removes AWS Regions in the **AWS Explorer**.                                                                                                                                                                                       |
 | `AWS: Sign out`                           | Disconnects the Toolkit from the currently-connected AWS account.                                                                                                                                                                          |
 | `AWS: Submit Quick Feedback...`           | Submit a private, one-way message and sentiment to the AWS Toolkit dev team. For larger issues that warrant conversations or bugfixes, please submit an issue in Github with the **AWS: Create a New Issue on Github** command.            |
 | `AWS: Toggle SAM hints in source files`   | Toggles AWS SAM-related Inline Actions in source files                                                                                                                                                                                     |
