+++ README.quickstart.vscode.md
@@ -224,13 +224,12 @@ palette_, available by selecting _View > Command Palette_ or by typing
 | `AWS: Detect SAM CLI`                                | Checks whether the Toolkit can communicate correctly with the AWS SAM CLI that is installed.                                                                                                                                                                                                                   |
 | `AWS: Edit SAM Debug Configuration`                  | Shows a tool that helps you create, edit, run, and debug a SAM _launch config_ (`type:aws-sam`).                                                                                                                                                                                                               |
 | `AWS: Focus on Explorer View`                        | Opens the **AWS** panel and focuses the [_Explorer_](#ui-components-aws-expl).                                                                                                                                                                                                                                 |
-| `AWS: Hide region from the Explorer`                 | Hides an AWS Region from the **AWS Explorer**.                                                                                                                                                                                                                                                                 |
 | `AWS: Publish state machine to Step Functions`       | Creates or updates a remote state machine using the local Amazon States Language definition file. For more information, see [Work With State Machines in VS Code](https://docs.aws.amazon.com/toolkit-for-vscode/latest/userguide/bulding-stepfunctions.html#starting-stepfunctions) in the user guide.        |
 | `AWS: Publish a Systems Manager Document`            | Creates a new document or updates an existing document in your account using the currently open document.                                                                                                                                                                                                      |
 | `AWS: Render state machine graph`                    | Renders the state machine definition into a graph visualization. For more information, see [State Machine Graph Visualization](https://docs.aws.amazon.com//toolkit-for-vscode/latest/userguide/bulding-stepfunctions.html#bulding-stepfunctions-visualizations) in the user guide.                            |
 | `AWS: Report an Issue`                               | In the [GitHub repository](https://github.com/aws/aws-toolkit-vscode) for the Toolkit, opens the page to [create a new issue](https://github.com/aws/aws-toolkit-vscode/issues/new/choose).                                                                                                                    |
 | `AWS: Save Current Log Content To File`              | Saves currently-loaded CloudWatch Log Stream data to a local `.log` file.                                                                                                                                                                                                                                      |
-| `AWS: Show region in the Explorer`                   | Displays an AWS Region in the **AWS Explorer**.                                                                                                                                                                                                                                                                |
+| `AWS: Show or Hide Regions`                          | Adds or removes AWS Regions in the **AWS Explorer**.                                                                                                                                                                                                                                                           |
 | `AWS: Sign out`                                      | Disconnects the Toolkit from the currently-connected AWS account.                                                                                                                                                                                                                                              |
 | `AWS: Submit Quick Feedback...`                      | Submit a private, one-way message and sentiment to the AWS Toolkit dev team. For larger issues that warrant conversations or bugfixes, please submit an issue in Github with the **AWS: Create a New Issue on Github** command.                                                                                |
 | `AWS: Toggle SAM hints in source files`              | Toggles AWS SAM-related Codelenses in source files                                                                                                                                                                                                                                                             |
