+++ package.json
@@ -43,7 +43,6 @@
         "onCommand:aws.createIssueOnGitHub",
         "onCommand:aws.submitFeedback",
         "onCommand:aws.showRegion",
-        "onCommand:aws.hideRegion",
         "onView:aws.explorer",
         "onCommand:aws.deploySamApplication",
         "onCommand:aws.samcli.detect",
@@ -1081,19 +1080,14 @@
                     "group": "1_account@1"
                 },
                 {
-                    "command": "aws.listCommands",
+                    "command": "aws.showRegion",
                     "when": "view == aws.explorer",
                     "group": "1_account@2"
                 },
                 {
-                    "command": "aws.showRegion",
+                    "command": "aws.listCommands",
                     "when": "view == aws.explorer",
-                    "group": "2_region@1"
-                },
-                {
-                    "command": "aws.hideRegion",
-                    "when": "view == aws.explorer",
-                    "group": "2_region@2"
+                    "group": "1_account@3"
                 },
                 {
                     "command": "aws.lambda.createNewSamApp",
@@ -1235,19 +1229,24 @@
                     "group": "inline@2"
                 },
                 {
-                    "command": "aws.lambda.createNewSamApp",
-                    "when": "view == aws.explorer && viewItem == awsLambdaNode",
+                    "command": "aws.showRegion",
+                    "when": "view == aws.explorer && viewItem == awsRegionNode",
                     "group": "0@1"
+                },
+                {
+                    "command": "aws.lambda.createNewSamApp",
+                    "when": "view == aws.explorer && viewItem == awsLambdaNode || viewItem == awsRegionNode",
+                    "group": "1@1"
                 },
                 {
                     "command": "aws.launchConfigForm",
                     "when": "view == aws.explorer && viewItem == awsLambdaNode || viewItem == awsRegionNode || viewItem == awsCloudFormationRootNode",
-                    "group": "0@2"
+                    "group": "1@1"
                 },
                 {
                     "command": "aws.deploySamApplication",
                     "when": "view == aws.explorer && viewItem == awsLambdaNode || viewItem == awsRegionNode || viewItem == awsCloudFormationRootNode",
-                    "group": "1@1"
+                    "group": "1@2"
                 },
                 {
                     "command": "aws.ecr.copyTagUri",
@@ -1498,11 +1497,6 @@
                     "command": "aws.downloadSchemaItemCode",
                     "when": "view == aws.explorer && viewItem == awsSchemaItemNode && !isCloud9",
                     "group": "1@1"
-                },
-                {
-                    "command": "aws.hideRegion",
-                    "group": "0@1",
-                    "when": "view == aws.explorer && viewItem == awsRegionNode"
                 },
                 {
                     "command": "aws.cloudWatchLogs.viewLogStream",
@@ -1796,16 +1790,6 @@
             {
                 "command": "aws.showRegion",
                 "title": "%AWS.command.showRegion%",
-                "category": "%AWS.title%",
-                "cloud9": {
-                    "cn": {
-                        "category": "%AWS.title.cn%"
-                    }
-                }
-            },
-            {
-                "command": "aws.hideRegion",
-                "title": "%AWS.command.hideRegion%",
                 "category": "%AWS.title%",
                 "cloud9": {
                     "cn": {
@@ -2900,8 +2884,7 @@
                             "markdown": "resources/walkthrough/setup-region.md"
                         },
                         "completionEvents": [
-                            "onCommand:aws.showRegion",
-                            "onCommand:aws.hideRegion"
+                            "onCommand:aws.showRegion"
                         ]
                     },
                     {
