+++ package.nls.json
@@ -160,8 +160,7 @@
     "AWS.command.createIssueOnGitHub": "Report Toolkit Issue",
     "AWS.command.createNewSamApp": "Create Lambda SAM Application",
     "AWS.command.credential.profile.create": "Create Credentials Profile",
-    "AWS.command.showRegion": "Show region in the Explorer",
-    "AWS.command.hideRegion": "Hide region from the Explorer",
+    "AWS.command.showRegion": "Show or Hide Regions",
     "AWS.command.deleteLambda.confirm": "Are you sure you want to delete lambda function '{0}'?",
     "AWS.command.deleteLambda.error": "There was an error deleting lambda function '{0}'",
     "AWS.command.addSamDebugConfiguration": "Add Debug Configuration",
@@ -275,8 +274,6 @@
     "AWS.error.mfa.userCancelled": "User cancelled entering authentication code",
     "AWS.error.no.error.code": "No error code",
     "AWS.error.sam.local.package_json_not_found": "Unable to find package.json related to {0}",
-    "AWS.explorerNode.addRegion": "Add a region to {0} Explorer...",
-    "AWS.explorerNode.addRegion.tooltip": "Click here to add a region to {0} Explorer.",
     "AWS.explorerNode.apigateway.error": "Error loading API Gateway REST APIs",
     "AWS.explorerNode.apigateway.noRestApis": "[No API Gateway REST APIs found]",
     "AWS.explorerNode.lambda.noFunctions": "[No Functions found]",
@@ -423,8 +420,6 @@
     "AWS.message.prompt.selectCDKWorkspace.placeholder": "Select CDK workspace",
     "AWS.message.prompt.selectCDKApplication.placeholder": "Select CDK application",
     "AWS.message.prompt.selectCDKStateMachine.placeholder": "Select a State Machine",
-    "AWS.message.prompt.region.hide.title": "Select a region to hide from the {0} Explorer",
-    "AWS.message.prompt.region.show.title": "Select a region to show in the {0} Explorer",
     "AWS.message.prompt.copyButtonLabel": "Copy",
     "AWS.message.prompt.changeS3RegionButton": "Change default S3 region",
     "AWS.message.toolkitInfo": "OS:  {0} {1} {2}\n{3} Extension Host Version:  {4}\n{5} Toolkit Version:  {6}\n",
