+++ src/extension.ts
@@ -28,7 +28,7 @@ import {
     githubUrl,
 } from './shared/constants'
 import { DefaultAwsContext } from './shared/awsContext'
-import { DefaultAWSContextCommands } from './shared/defaultAwsContextCommands'
+import { AwsContextCommands } from './shared/awsContextCommands'
 import {
     aboutToolkit,
     getIdeProperties,
@@ -49,14 +49,7 @@ import { activate as activateSam } from './shared/sam/activation'
 import { DefaultSettingsConfiguration } from './shared/settingsConfiguration'
 import { activate as activateTelemetry } from './shared/telemetry/activation'
 import { activate as activateS3 } from './s3/activation'
-import {
-    recordAwsCreateCredentials,
-    recordAwsHelp,
-    recordAwsHelpQuickstart,
-    recordAwsReportPluginIssue,
-    recordAwsShowExtensionSource,
-    recordToolkitInit,
-} from './shared/telemetry/telemetry'
+import * as telemetry from './shared/telemetry/telemetry'
 import { ExtContext } from './shared/extensions'
 import { activate as activateApiGateway } from './apigateway/activation'
 import { activate as activateStepFunctions } from './stepFunctions/activation'
@@ -119,12 +112,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
         await initializeAwsCredentialsStatusBarItem(awsContext, context)
         globals.regionProvider = regionProvider
-        globals.awsContextCommands = new DefaultAWSContextCommands(
-            awsContext,
-            awsContextTrees,
-            regionProvider,
-            loginManager
-        )
+        globals.awsContextCommands = new AwsContextCommands(awsContext, awsContextTrees, regionProvider, loginManager)
         globals.sdkClientBuilder = new DefaultAWSClientBuilder(awsContext)
         globals.toolkitClientBuilder = new DefaultToolkitClientBuilder(regionProvider)
         globals.schemaService = new SchemaService(context)
@@ -172,7 +160,7 @@ export async function activate(context: vscode.ExtensionContext) {
                 try {
                     await globals.awsContextCommands.onCommandCreateCredentialsProfile()
                 } finally {
-                    recordAwsCreateCredentials()
+                    telemetry.recordAwsCreateCredentials()
                 }
             })
         )
@@ -181,19 +169,19 @@ export async function activate(context: vscode.ExtensionContext) {
         context.subscriptions.push(
             vscode.commands.registerCommand('aws.help', async () => {
                 vscode.env.openExternal(vscode.Uri.parse(documentationUrl))
-                recordAwsHelp()
+                telemetry.recordAwsHelp()
             })
         )
         context.subscriptions.push(
             vscode.commands.registerCommand('aws.github', async () => {
                 vscode.env.openExternal(vscode.Uri.parse(githubUrl))
-                recordAwsShowExtensionSource()
+                telemetry.recordAwsShowExtensionSource()
             })
         )
         context.subscriptions.push(
             vscode.commands.registerCommand('aws.createIssueOnGitHub', async () => {
                 vscode.env.openExternal(vscode.Uri.parse(githubCreateIssueUrl))
-                recordAwsReportPluginIssue()
+                telemetry.recordAwsReportPluginIssue()
             })
         )
         context.subscriptions.push(
@@ -201,7 +189,7 @@ export async function activate(context: vscode.ExtensionContext) {
                 try {
                     await showQuickStartWebview(context)
                 } finally {
-                    recordAwsHelpQuickstart({ result: 'Succeeded' })
+                    telemetry.recordAwsHelpQuickstart({ result: 'Succeeded' })
                 }
             })
         )
@@ -336,7 +324,7 @@ function recordToolkitInitialization(activationStartedOn: number, logger?: Logge
         const activationFinishedOn = Date.now()
         const duration = activationFinishedOn - activationStartedOn
 
-        recordToolkitInit({
+        telemetry.recordToolkitInit({
             duration: duration,
         })
     } catch (err) {
